My_first_store/requirements.txt
Django==5.2.7
djangorestframework==3.15.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==22.0.0